requirement.txt
# Mistral AI
mistralai>=1.0.0
langchain-mistralai>=0.1.0

# LangChain core
langchain>=0.2.0
langchain-core>=0.2.0

# Environment variable management
python-dotenv>=1.0.0

# Optional but recommended utilities
requests>=2.31.0
httpx>=0.27.0